% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.data.frame.murphydiag}
\alias{as.data.frame.murphydiag}
\title{Coerce a Murphy diagram to a Data Frame}
\usage{
\method{as.data.frame}{murphydiag}(x, ...)
}
\arguments{
\item{x}{\code{'murphydiag'} object to be coerced to \code{'data.frame'}}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
Returns a \code{'data.frame'} with four columns:
 \code{name}, \code{knot}, \code{limit}, \code{value}
}
\description{
Coerce a Murphy diagram to a Data Frame
}
\details{
Coercion of a \code{'murphydiag'} object to a \code{'data.frame'}
 retains only the relevant information for plotting: the model/forecast
 name, the knots (break points) of the piecewise linear functions,
 and the values taken in the left and right limits in the knots.
 
 Additional information such as the observations, the predictions,
 and the functional type of the predictions is lost. Hence, reversion
 to \code{'murphydiag'} is impossible. When coercing a
 \code{'data.frame'} to a \code{'murphydiag'}, it is assumed that the
 \code{'data.frame'} comprises predictions (from multiple methods).
}
