% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphydiag_dom.R
\name{dominance}
\alias{dominance}
\title{dominance}
\usage{
dominance(m, abs.tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{m}{An object inheriting from the "murphydiag" class.}

\item{abs.tol}{absolute tolerance}
}
\value{
A matrix representing the dominance relationships (partial order)
  of the forecast methods in \code{m}.
}
\description{
dominance
}
