% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{drawExample}
\alias{drawExample}
\title{Examples}
\usage{
drawExample(n = 500, m = 30)
}
\arguments{
\item{n}{Number of forecast observation pairs.}

\item{m}{Number of forecast methods.}
}
\value{
A list.
}
\description{
Examples
}
