% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementaryscores.R
\name{elementaryscores}
\alias{elementaryscores}
\alias{es_mean}
\alias{es_prob}
\alias{es_median}
\alias{es_quant}
\alias{es_expect}
\title{Elementary score functions}
\usage{
es_mean(t, x, y, right = FALSE)

es_prob(t, x, y, right = FALSE)

es_median(t, x, y, right = FALSE)

es_quant(t, x, y, level, right = FALSE)

es_expect(t, x, y, level, right = FALSE)
}
\arguments{
\item{t}{a single threshold value.}

\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations,
#forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{right}{logical, indicating if the scores are
right-continuous in \code{x}.}

\item{level}{a single value in (0, 1).}
}
\value{
yadda
}
\description{
Functions for the computation of the elementary scores.
}
