% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{elementaryscores_fun}
\alias{elementaryscores_fun}
\alias{es_expect_fun}
\alias{es_mean_fun}
\alias{es_median_fun}
\alias{es_prob_fun}
\alias{es_quant_fun}
\title{Elementary score function constructors}
\usage{
es_expect_fun(x, y, level)

es_mean_fun(x, y)

es_median_fun(x, y)

es_prob_fun(x, y)

es_quant_fun(x, y, level)
}
\arguments{
\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations,
#forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{level}{numerical, a single value between 0 and 1.}
}
\value{
A function in two arguments: 
  \tabular{ll}{
  \code{t} \tab an object convertible by \code{as.vector}
 containing treshold values.\cr
  \code{right} \tab logical, indicating if the scores are
  right-continuous in \code{t}.
  }
}
\description{
Creating functions for the computation of mean elementary scores.
}
