% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots.R
\name{knots.murphydiag}
\alias{knots.murphydiag}
\title{Getting the 'knots' from a "murphydiag" object}
\usage{
\method{knots}{murphydiag}(Fn, ...)
}
\arguments{
\item{Fn}{an R object inheriting from "murphydiag"}

\item{...}{aditional parameters.}
}
\value{
A list of knot vectors; one for each forecasting method in \code{Fn}.
}
\description{
Getting the 'knots' from a "murphydiag" object
}
