% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m_filter}
\alias{m_filter}
\alias{m_approx}
\title{Data frame approximations of Murphy diagrams}
\usage{
m_filter(m, window = range(m), resolution = 100)

m_approx(m, thresholds = 100)
}
\arguments{
\item{m}{an object inheriting from class \code{'murphydiag'}}

\item{window}{a numeric vector of the form \code{c(x1, x2, y1, y2)};
e.g., \code{range(m)} or \code{par("usr")}}

\item{resolution}{a numeric vector of the form \code{c(nx, ny)}
by which to divide \code{window}}

\item{thresholds}{a numeric vector of values where interpolation is to take place;
alternatively, a single number specifying the amount of
equally spaced points spanning \code{range(m, dim = "knots")}}
}
\value{
\code{m_filter} returns a \code{data.frame} with four columns: \code{name}, \code{knot}, \code{limit}, \code{value}. The returned \code{data.frame} comprises a subset of the rows returned by \code{as.data.frame(m)}.
 
 \code{m_approx} returns a \code{data.frame} with three columns: \code{name}, \code{threshold}, \code{value}
}
\description{
Data frame approximations of Murphy diagrams
}
\details{
The primary purpose of \code{m_filter} and \code{m_approx}
 is to reduce the amount of points in \code{plot(m)}.
 
 \code{m_filter} aims to retain discontinuities and
 a visual impression that is close to the true empirical Murphy diagram.
 
 \code{m_approx} performs smooth linear interpolation between midpoints,
 evaluated at given threshold values on the x-axis.
 
 This is achieved by two different methods:
 
 \code{m_filter} moves sequentially through the rows
 of \code{as.data.frame(m)}. Any point with a value close to 0
 on the y-axis is retained. Otherwise,
 a rectangle with dimensions \code{c(x2 - x1, y2 - y1) / c(nx, ny)}
 is placed, with the latest retained point at its center.
 The next point that falls outside of this rectangle
 (and that point's predecessor) are then retained,
 and any points inbetween are discarded.
 
 \code{m_approx} performs a linear approximation
 on the data set of the knots and the midpoints
 of left-sided and right-sided limits
 of the Murphy diagram values in the respective knots.
 Afterwards that approximation is evaluated
 at user-specified thresholds or
 at a user-specified number of equally spaced points
 spanning \code{range(m, dim = "knots")}.
}
\seealso{
For an exact \code{data.frame} representation see \code{\link{as.data.frame.murphydiag}}.
 
 \code{\link{range.murphydiag}}
}
