% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanscores_fun.R
\name{meanscores_fun}
\alias{meanscores_fun}
\alias{ms_fun}
\alias{ms_mean_fun}
\alias{ms_prob_fun}
\alias{ms_median_fun}
\alias{ms_quant_fun}
\alias{ms_expect_fun}
\title{Mean score function constructors}
\usage{
ms_fun(m)

ms_mean_fun(x, y)

ms_prob_fun(x, y)

ms_median_fun(x, y)

ms_quant_fun(x, y, level)

ms_expect_fun(x, y, level)
}
\arguments{
\item{m}{an object inheriting from class "murphydiag".}

\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations, #forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{level}{a single value in (0, 1).}
}
\value{
hubba
}
\description{
Creating functions for the computation of mean elementary scores.
}
