% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphydiag_diff.R
\name{murphydiag_diff}
\alias{murphydiag_diff}
\alias{-.murphydiag}
\title{Murphy diagram difference object}
\usage{
murphydiag_diff(m1, m2 = NULL)

\method{-}{murphydiag}(m1, m2)
}
\arguments{
\item{m1}{an object inheriting from class \code{murphydiag}.}

\item{m2}{optional; an object that can be combined
into an existing \code{murphydiag} object
(see \link{c.murphydiag}).}
}
\value{
returns a \code{murphydiag_diff} object to be used
  with \code{plot} (see \link{plot.murphydiag_diff})
}
\description{
\code{murphydiag_diff} constructors
}
\section{Methods (by generic)}{
\itemize{
\item \code{-}: 
}}

\seealso{
\code{\link{murphydiag}},
  \code{\link{c.murphydiag}},
  \code{\link{[.murphydiag}},
  \code{\link{plot.murphydiag_diff}}
}
