% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dominance}
\alias{plot.dominance}
\title{Plot dominance objects}
\usage{
\method{plot}{dominance}(x, ..., select = NULL, origin = NULL)
}
\arguments{
\item{x}{an object inheriting from the class \code{dominance}.}

\item{...}{additional parameters passed to \code{hasseDiagram::hasse}.}

\item{select}{choose a subset of forecasting methods.}

\item{origin}{choose a set of origins.}
}
\description{
Plot dominance objects
}
