% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.murphydiag}
\alias{plot.murphydiag}
\alias{lines.murphydiag}
\title{Plotting Murphy diagram objects}
\usage{
\method{plot}{murphydiag}(x, resolution = 100, xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, ..., add = FALSE,
  no_baseline = FALSE, empty = FALSE, lty = 1:5, col = 1:6)

\method{lines}{murphydiag}(x, resolution = 100, ..., lty = 1:5,
  col = 1:6)
}
\arguments{
\item{x}{an object inheriting from the class \code{'murphydiag'}.}

\item{resolution}{determines the smoothness (and accuracy) of the displayed lines:
\enumerate{
\item if \code{NULL}: all points returned by \code{as.data.frame(x)} are displayed.
\item if \code{c(n)} or \code{c(nx, ny)}: only points returned by \code{m_filter(x, window, resolution)} with \code{window = par("usr")} are displayed.
\item if \code{c(t1, t2, ..., tn)}: only points returned by \code{m_approx(x, thresholds)} with \code{thresholds = resolution} are displayed.
}
See \code{\link[as.data.frame.murphydiag]{as.data.frame}}, \code{\link{m_filter}}, \code{\link{m_approx}}.}

\item{xlim}{the x limits (x1, x2) of the plot.  Note that \code{x1 > x2}
    is allowed and leads to a \sQuote{reversed axis}.

    The default value, \code{NULL}, indicates that the range of the
    \link{finite} values to be plotted should be used.}

\item{ylim}{the y limits of the plot.}

\item{xlab}{a label for the x axis, defaults to "threshold".}

\item{ylab}{a label for the y axis, defaults to "mean score".}

\item{...}{other graphical parameters (see \code{\link{par}}).}

\item{add}{if \code{TRUE}, lines are added to the current plot.}

\item{no_baseline}{if \code{TRUE}, no horizontal baseline at 0 is drawn.}

\item{empty}{if \code{TRUE}, no Murphy diagram lines are drawn.}

\item{lty}{vector of line types, recycled to \code{length(x)}.}

\item{col}{vector of colors, recycled to \code{length(x)}.}
}
\description{
Plotting Murphy diagram objects
}
\seealso{
\code{\link{murphydiag}}.
  
  To display only a selection of Murphy diagram lines,
  consider subsetting (see \code{\link{[.murphydiag}}).
}
