% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.murphydiag_diff}
\alias{plot.murphydiag_diff}
\title{Plotting Murphy diagram difference objects}
\usage{
\method{plot}{murphydiag_diff}(x, ..., thresholds = 500L, type = "l",
  xlim = NULL, ylim = NULL, main = NULL, xlab = NULL,
  ylab = NULL, level_ci = 0.95, sd_lags = 0)
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag_diff}.}

\item{thresholds}{either the maximum number of thresholds (if length 1) or the values against which to plot}

\item{type}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}

\item{xlim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{ylim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{main}{an overall title for the plot: see \code{\link{title}}.}

\item{xlab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{ylab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{level_ci}{level for confidence interval (set to \code{NULL} to omit it)}

\item{sd_lags}{number of autocovariances to use for confidence intervals
(defaults to zero; values 1, 2, 3, ... may make sense for time series data.
 Set to \code{NULL} for data-driven selection.)}
}
\description{
Plotting Murphy diagram difference objects
}
\seealso{
\code{\link{murphydiag_diff}},
  \code{\link{-.murphydiag}},
  \code{\link{[.murphydiag}}
}
