% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{range.murphydiag}
\alias{range.murphydiag}
\title{Range of values in Murphy diagrams}
\usage{
\method{range}{murphydiag}(..., dim = c("knots", "values"),
  na.rm = FALSE)
}
\arguments{
\item{...}{one or multiple objects inheriting from the class \code{'murphydiag'}.}

\item{dim}{a character vector that specifies
along which dimensions the range should be computed.}

\item{na.rm}{for compatibility with the S3 generic function.}
}
\value{
a numeric vector of the form \code{c(k1, k2, v1, v2)},
 where
 \enumerate{
 \item the smallest knot \code{k1} and largest knot \code{k2}
 are dropped if \code{dim} does not contain the string \code{"knots"},
 \item the smallest value \code{v1} and largest value \code{v2}
 are dropped if \code{dim} does not contain the string \code{"values"}.
 }
}
\description{
Retrieve the range of knots and/or values
across all entries of one or multiple \code{'murphydiag'} objects.
Other objects are coerced by \code{as.murphydiag}.
}
