% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullscores.R, R/meanscores.R
\name{fs_mean}
\alias{fs_mean}
\alias{fs_prob}
\alias{fs_median}
\alias{fs_quant}
\alias{fs_expect}
\alias{scores}
\alias{ms_mean,}
\alias{ms_prob,}
\alias{ms_median,}
\alias{ms_quant,}
\alias{ms_expect,}
\alias{fs_mean,}
\alias{fs_prob,}
\alias{fs_median,}
\alias{fs_quant,}
\alias{fs_expect}
\alias{ms_mean}
\alias{ms_prob}
\alias{ms_median}
\alias{ms_quant}
\alias{ms_expect}
\title{Score functions}
\usage{
fs_mean(t, x, y, right = FALSE)

fs_prob(t, x, y, right = FALSE)

fs_median(t, x, y, right = FALSE)

fs_quant(t, x, y, level, right = FALSE)

fs_expect(t, x, y, level, right = FALSE)

ms_mean(t, x, y, right = FALSE)

ms_prob(t, x, y, right = FALSE)

ms_median(t, x, y, right = FALSE)

ms_quant(t, x, y, level, right = FALSE)

ms_expect(t, x, y, level, right = FALSE)
}
\arguments{
\item{t}{an object convertible by \code{as.vector}
containing threshold values.}

\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations, #forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{right}{logical, indicating if the scores are
right-continuous in \code{x}, hence
left-continuous in \code{t}.}

\item{level}{a single value in (0, 1).}
}
\value{
\code{\link{ms_mean}} and all other functions starting with \code{ms_} 
  return a matrix whose rows correspond to values for \code{t}, and whose cols correspond to forecast methods. The cells of the matrix are average elementary scores over forecast cases. 
  \code{\link{fs_mean}} and all other functions starting with \code{fs_} return a list whose entries correspond to values for \code{t}. Each entry contains a matrix whose rows correspond to forecast cases, and whose columns correspond to forecast methods.
}
\description{
Functions for the computation of elementary scores.
}
