% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullscores_fun.R, R/meanscores_fun.R
\name{fs_fun}
\alias{fs_fun}
\alias{fs_mean_fun}
\alias{fs_prob_fun}
\alias{fs_median_fun}
\alias{fs_quant_fun}
\alias{fs_expect_fun}
\alias{scores_fun}
\alias{ms_fun,}
\alias{ms_mean_fun,}
\alias{ms_prob_fun,}
\alias{ms_median_fun,}
\alias{ms_quant_fun,}
\alias{ms_expect_fun,}
\alias{fs_fun,}
\alias{fs_mean_fun,}
\alias{fs_prob_fun,}
\alias{fs_median_fun,}
\alias{fs_quant_fun,}
\alias{fs_expect_fun}
\alias{ms_fun}
\alias{ms_mean_fun}
\alias{ms_prob_fun}
\alias{ms_median_fun}
\alias{ms_quant_fun}
\alias{ms_expect_fun}
\title{Score function constructors}
\usage{
fs_fun(m)

fs_mean_fun(x, y)

fs_prob_fun(x, y)

fs_median_fun(x, y)

fs_quant_fun(x, y, level)

fs_expect_fun(x, y, level)

ms_fun(m)

ms_mean_fun(x, y)

ms_prob_fun(x, y)

ms_median_fun(x, y)

ms_quant_fun(x, y, level)

ms_expect_fun(x, y, level)
}
\arguments{
\item{m}{an object inheriting from class "murphydiag".}

\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations, #forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{level}{a single value in (0, 1).}
}
\value{
Function which gives the elementary score as a function of the threshold parameter \code{t}.
}
\description{
Creating functions for the computation of elementary scores.
}
