% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extract.R
\name{[.dominance}
\alias{[.dominance}
\title{Subsetting dominance objects}
\usage{
\method{[}{dominance}(d, j)
}
\arguments{
\item{d}{an object inheriting from the class \code{dominance}.}

\item{j}{index specifying elements to extract.}
}
\value{
(see \link{[.data.frame}).
}
\description{
Subsetting dominance objects
}
