% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_murphydiag.R
\name{[.murphydiag}
\alias{[.murphydiag}
\title{Subsetting Murphy diagram objects}
\usage{
\method{[}{murphydiag}(m, j)
}
\arguments{
\item{m}{an object inheriting from the class \code{murphydiag}.}

\item{j}{index specifying elements to extract.}
}
\value{
(see \link{[.data.frame}).
}
\description{
And split 'em back up
}
\seealso{
\code{\link{c.murphydiag}}
}
