% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.data.frame.murphydiag}
\alias{as.data.frame.murphydiag}
\title{Converting a murphydiag object into a data.frame}
\usage{
\method{as.data.frame}{murphydiag}(x, ..., thresholds = 10000L,
  right.cont = "both")
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag}.}

\item{thresholds}{either the maximum number of thresholds (if length 1) or the values against which to plot}

\item{right.cont}{logical; whether the diagrams are right-continuous or not. Alternatively: the string "both"}
}
\value{
A \code{data.frame} with three columns: \code{method}, \code{threshold}, \code{score}
}
\description{
Converting a murphydiag object into a data.frame
}
