% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_murphydiag.R
\name{as.data.frame.murphydiag}
\alias{as.data.frame.murphydiag}
\title{Converting a murphydiag object into a data.frame}
\usage{
\method{as.data.frame}{murphydiag}(x, ..., thresholds = 500L)
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag}.}

\item{...}{Graphical parameters (see \code{\link{par}}) and any further
    arguments of \code{plot}, typically \code{\link{plot.default}}, may also be
    supplied as arguments to this function.  Hence, the high-level
    graphics control arguments described under \code{\link{par}} and the
    arguments to \code{\link{title}} may be supplied to this function.}

\item{thresholds}{either the maximum number of thresholds (if length 1) or the values against which to plot}
}
\value{
A \code{data.frame} with three columns: \code{method}, \code{threshold}, \code{score}
}
\description{
Converting a murphydiag object into a data.frame
}
