% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_murphydiag.R
\name{c.murphydiag}
\alias{c.murphydiag}
\title{Combining Murphy diagram objects}
\usage{
\method{c}{murphydiag}(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\value{
An object inheriting from the \code{murphydiag} class.
}
\description{
Combine two or more 'murphydiag' objects that are based on the same observations and type of forecasts.
}
\seealso{
\code{\link{[.murphydiag}}, \code{\link{murphydiag}}
}
