% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{meanscores_fun}
\alias{meanscores_fun}
\alias{ms_mean_fun}
\alias{ms_prob_fun}
\alias{ms_median_fun}
\alias{ms_quant_fun}
\alias{ms_expect_fun}
\title{Mean score function constructors}
\usage{
ms_mean_fun(x, y)

ms_prob_fun(x, y)

ms_median_fun(x, y)

ms_quant_fun(x, y, level)

ms_expect_fun(x, y, level)
}
\arguments{
\item{x}{an object convertible by \code{as.matrix}
containing forecasts (dim: #observations,
#forecast methods).}

\item{y}{an object convertible by \code{as.vector}
containing observations.}

\item{level}{numerical, a single value between 0 and 1.}
}
\value{
A function in two arguments: 
  \tabular{ll}{
  \code{t} \tab an object convertible by \code{as.vector}
 containing treshold values.\cr
  \code{right} \tab logical, indicating if the scores are
  right-continuous in \code{t}.
  }
}
\description{
Creating functions for the computation of mean elementary scores.
}
