% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_murphydiag.R
\name{plot.murphydiag}
\alias{plot.murphydiag}
\title{Plotting Murphy diagram objects}
\usage{
\method{plot}{murphydiag}(x, thresholds = 500L, type = "l", xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag}.}

\item{thresholds}{either the maximum number of thresholds (if length 1) or the values against which to plot}

\item{type}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}

\item{xlim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{ylim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{xlab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{ylab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{...}{Graphical parameters (see \code{\link{par}}) and any further
    arguments of \code{plot}, typically \code{\link{plot.default}}, may also be
    supplied as arguments to this function.  Hence, the high-level
    graphics control arguments described under \code{\link{par}} and the
    arguments to \code{\link{title}} may be supplied to this function.}
}
\description{
Plotting Murphy diagram objects
}
\seealso{
\code{\link{murphydiag}},
  \code{\link{c.murphydiag}},
  \code{\link{[.murphydiag}}
}
