% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.murphydiag}
\alias{plot.murphydiag}
\title{Plotting Murphy diagram objects}
\usage{
\method{plot}{murphydiag}(x, thresholds = 10000L, right.cont = "both",
  type = "l", xlim = NULL, ylim = NULL, xlab = NULL, ylab = NULL,
  ...)
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag}.}

\item{thresholds}{either the maximum number of thresholds (if length 1) or the values against which to plot}

\item{right.cont}{logical; whether the diagrams are right-continuous or not. Alternatively: the string "both"}

\item{type}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}

\item{xlim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{ylim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{xlab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{ylab}{titles for x and y axes, as in \code{\link{plot}}.}
}
\description{
Plotting Murphy diagram objects
}
\seealso{
\code{\link{murphydiag}},
  \code{\link{c.murphydiag}},
  \code{\link{[.murphydiag}}
}
