% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_murphydiag_diff.R
\name{plot.murphydiag_diff}
\alias{plot.murphydiag_diff}
\title{Plotting Murphy diagram difference objects}
\usage{
\method{plot}{murphydiag_diff}(x, type = "l", xlim = NULL, ylim = NULL,
  main = NULL, xlab = NULL, ylab = NULL, level_ci = 0.95, ...)
}
\arguments{
\item{x}{an object inheriting from the class \code{murphydiag}.}

\item{type}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}

\item{xlim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{ylim}{ranges of x and y axes, as in \code{\link{plot}}.}

\item{main}{an overall title for the plot: see \code{\link{title}}.}

\item{xlab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{ylab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{level_ci}{level for confidence interval}

\item{...}{Graphical parameters (see \code{\link{par}}) and any further
    arguments of \code{plot}, typically \code{\link{plot.default}}, may also be
    supplied as arguments to this function.  Hence, the high-level
    graphics control arguments described under \code{\link{par}} and the
    arguments to \code{\link{title}} may be supplied to this function.}
}
\description{
Plaster them walls
}
\seealso{
\code{\link{murphydiag_diff}},
  \code{\link{-.murphydiag}},
  \code{\link{[.murphydiag}}
}
