% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extract.R
\name{[.murphydiag}
\alias{[.murphydiag}
\title{Subsetting Murphy diagram objects}
\usage{
\method{[}{murphydiag}(x, i)
}
\arguments{
\item{x}{an object inheriting from the class \code{'murphydiag'}.}

\item{i}{index specifying which elements to extract.}
}
\value{
an object inheriting from the class \code{'murphydiag'}.
}
\description{
Subsetting Murphy diagram objects
}
\seealso{
\code{\link{c.murphydiag}}.
}
