% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extract.R
\name{[.murphydiag}
\alias{[.murphydiag}
\title{Subsetting Murphy diagram objects}
\usage{
\method{[}{murphydiag}(m, i)
}
\arguments{
\item{m}{an object inheriting from the class \code{murphydiag}.}

\item{i}{index specifying elements to extract.}
}
\value{
(see \link{[.data.frame}).
}
\description{
Subsetting Murphy diagram objects
}
\seealso{
\code{\link{c.murphydiag}}
}
